requirements.txt
aniso8601==9.0.1
blinker==1.6.2
camelot-py==0.11.0
cffi==1.15.1
chardet==5.1.0
charset-normalizer==3.1.0
click==8.1.3
cryptography==41.0.1
et-xmlfile==1.1.0
Flask==2.3.2
Flask-RESTful==0.3.10
ghostscript==0.7
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mysql-connector-python==8.1.0
numpy==1.25.0
opencv-python==4.7.0.72
openpyxl==3.1.2
pandas==2.0.2
pdfkit==1.0.0
pdfminer.six==20221105
protobuf==4.21.12
pycparser==2.21
PyJWT==2.8.0
pypdf==3.10.0
python-dateutil==2.8.2
pytz==2023.3
six==1.16.0
tabulate==0.9.0
tzdata==2023.3
Werkzeug==2.3.6
